Fix weird constraint positioning behaviour
For some reason when I don't include 'pointB' in
'bottomForearmConstraintArgs', even though its exactly
the same in 'topForearmConstraintArgs's when we copt it,
the spring constraints' fix point on the arm moves up and down,
when it should stay in the same place!

No idea why this change fixes this, but giving
'bottomForearmConstraintArgs' its own 'pointB' vs getting
the value from the spread, does the job.