Add basic mouse interactivity logic
This stuff needs cleaning up and module-ising, but this is basically
how we'll want to control the arm with the mouse.

For vertical movement, we apply a constant vertical force to the hand,
pushing and holding it upwards or downwards depending on the mouse's Y.
Matter.js resets bodies' force vectors with each engine tick, so we need
to make sure we set it every time "before update".

For horizontal movement, we manually set the 'elbow's position. All of
the arm, including the forearm springs, are fixed to the elbow. We have
to set the elbow's position, vs applying a force to it, because it's a
'static' body, i.e. it has infinite mass. I've made it static so it
isn't nudged or waggled by the motion of the arm.

Also increased the length of the forearm springs so they're further away
from the arm, making for looser forearm movement.