Fix elbow with constraints vs making static
Keeping the elbow as a static body, giving it infinite
mass, means when we add a second arm into the world,
it won't be able to be pushed back and nudged by the first.
But this will be important to the look and feel of the
'handshake' - we need to the two arms to respond to eachother.

Instead, the elbow is held in place using four new spring
constraints, with looser, longer horizontal constraints to
allow the arms to be pushed back by one another.

(Really need to cut down the Arm constructor, but it's a bit
tricky as so much of the bodies and constraint creation stuff
relies on many many variables in the constructor scope...)