Create Arm's world constraints wrt (0, 0) then translate
The Composite class' 'translate' function doesn't translate constraints
that are fixed to world-space, meaning we couldn't do what we do with
the arm, forearm, and elbow bodies, i.e. create them relative to (0, 0)
then translate.

But this change adds a 'translateCompositeWithConstraints' function in a
new 'Utils' module that does just that. Now, we can create all of Arm's
bodies and constraints relative to (0, 0), then translate them to
the position argument given to the constructor at the end - nice and
simple.