Param-ise ArmController's 'mouseArea'; scale bodies' pos and size to container div
Also general cleanup in main.js.

The 'mouse area' in which mouse movement affects the forces applied
to the arm in ArmController, is now configurable via a 'dimens'
parameter. This mouse area is centered in the canvas, so when the
mouse is at the canvas' centre point, no force will be applied
to the arm.

Also, main.js now expects a div with id 'handshake-container' to exist
in the document, which is given to the Render.create call, vs
just giving 'document.body'.