Keep arms at fixed size + add mouseAreaDimens update logic
Rather than having the canvas size be relative to the size of the
container, it's now created at a fixed 800x400. The arms' sizes are also
fixed, letting us avoid the issue of having to scale the forces we apply
to them with their size (matter.js makes a couple of physical properties
depdenent on a Body's size, making consistent arm movement across
smaller and larger arms difficult).

I've also added a bit of flex magic to the container's css which will
make sure the canvas is always centered in the window. When the canvas
is moved about the page though, the ArmController's mouse-position-to-force
maths is thrown off, so we need to make sure we're telling the
controllers whenever the canvas moves. The solution here was to turn the
ArmController stuff into a class, which offers a 'canvasDimensChanged'
method.