Add 'Hand' class
We need to add new bodies on the front, top, and bottom of the hand,
which will allow us to tell where the two hands make contact. I.e. we
can check for collisions between the two front 'bumper' bodies for a
front bump, the left arm's top bumper and the right arm's bottom bumper
for a top-bottom bump, etc.

Creating and constraining all of these bumper bodies took up a lot of
lines, so I've wrapped it all up in the new Hand class.

Using matter.js' 'collisionFilter' system allows us to ensure that only
the matching bumpers collide: the front contact can only collide with
the other front contact, the left arm's top bumper with the right hand's
bottom, etc.